requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
diskcache==5.6.3
distro==1.9.0
docstring_parser==0.17.0
fastapi==0.135.2
frozenlist==1.8.0
grpcio==1.80.0
grpcio-health-checking==1.80.0
grpcio-tools==1.80.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
instructor==1.14.5
Jinja2==3.1.6
jiter==0.11.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
multidict==6.7.1
openai==2.30.0
packaging==26.0
pluggy==1.6.0
propcache==0.4.1
protobuf==6.33.6
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
pytest==9.0.2
pytest-mock==3.15.1
python-dotenv==1.2.2
requests==2.33.1
rich==14.3.3
ruff==0.15.8
setuptools==82.0.1
shellingham==1.5.4
sniffio==1.3.1
starlette==1.0.0
tenacity==9.1.4
tqdm==4.67.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.42.0
yarl==1.23.0
